sqalloc starts looking very cool with the viz. merging still missing

+++ os.h
@@ -2,10 +2,11 @@
 #include <stdio.h> // for debugging, nothing more
 #include <stdlib.h>
 #include <string.h> // memcpy
+#include <math.h>
 
 typedef unsigned char uchar;
 typedef unsigned short u16int;
 typedef unsigned int u32int;
 typedef unsigned long long u64int;
 
-#define nelem(x) (sizeof(x)/sizeof(x[0]))
+#define nelem(x) (int)(sizeof(x)/sizeof(x[0]))
